h26: Move CompareSection.tsx into /compare/ dir
